fix: updated dimo sdk for content-type checking fix
The DIMO Python SDK was recently updated with more
robust, but stricted response checks in order to
provide correctly decoded response data.

This worked fine at the time when v0.7.0 of the HA
integration was published, but yesterday the DIMO API was
updated and now requests return a different content type.

This caused returned data to not be json decoded which is what
broke the Home Assistant integration

The DIMO Python SDK has now been updated with a refactored
content type check.

Fixes #185