refactor: implemented chunked signals retrieval
DIMO has recently intoduced a (conservative) complexity
limit for Graphql queries. For vehicles with a large number
of signals we end up exceeding this limit when we query
for the values of all signals.

To work around this issue we can chunk up the queries and
merge the responses in order to get all signal values

Fixes #199